requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
sqlalchemy
psycopg2
python-dotenv
passlib[bcrypt]
python-jose
supabase
paddlepaddle
paddleocr
opencv-python-headless
pymupdf
openai
